Added language to hatch constructor and extra fileds in manifest
Language is needed when packing so we need to pass it in durng hatch
creation and we now also put the hatch name and language in the
hatch_metadata file to be easily parsed by the feed processing pipeline.

Also removed now is the requirement to pass the argv array since we can
grab it directly and only handle it if overrides are needed.

+++ test/lib/index.test.js
@@ -9,37 +9,62 @@ describe('Hatch', function() {
     let hatch;
 
     afterEach(() => {
-        fs.rmdirSync(hatch.get_path());
+        if (fs.existsSync(hatch.get_path())) {
+            fs.rmdirSync(hatch.get_path());
+        }
     });
 
-    it('can return path of hatch', function() {
-        hatch = new libingester.Hatch("abcd");
-        expect(hatch.get_path()).to.match(/hatch_abcd_[0-9_]+/);
-    });
+    describe('required params', function() {
+        it('can return path of hatch', function() {
+            hatch = new libingester.Hatch("abcd", "en");
+            expect(hatch.get_path()).to.match(/hatch_abcd_[0-9_]+/);
+        });
 
-    it('can be forced to use a specific path', function() {
-        hatch = new libingester.Hatch("abcd", { path: "./foo_bar_baz" });
-        expect(hatch.get_path()).to.match(/foo_bar_baz/);
+        it('can return the name of hatch', function() {
+            hatch = new libingester.Hatch("testing", "en");
+            expect(hatch.get_name()).to.equal("testing");
+        });
+
+        it('can return the language of hatch', function() {
+            hatch = new libingester.Hatch("abcd", "something");
+            expect(hatch.get_language()).to.equal("something");
+        });
+
+        it('can be forced to use specific path', function() {
+            hatch = new libingester.Hatch("abcd", "en", { path: "./foo_bar_baz" });
+            expect(hatch.get_path()).to.match(/foo_bar_baz/);
+        });
+
+        it('requires name and lang parameters to instantiate', function() {
+            expect(() => { new libingester.Hatch() }).to.throw();
+            expect(() => { new libingester.Hatch("abcd") }).to.throw();
+        });
+
+        // XXX: This is to ensure that v2-converted ingesters know that they
+        //      need to use the newer api.
+        it('requires second param to be a string', function() {
+            expect(() => { new libingester.Hatch("abcd", { foo: "bar" }) }).to.throw();
+        });
     });
 
     describe('argv no-tgz option', function() {
         it('does not blow up when no-tgz arg is missing', function() {
             // Implicit non-exception
-            hatch = new libingester.Hatch("aacd", { argv: ["--tgz", "/some/path"] });
+            hatch = new libingester.Hatch("aacd", "en", { argv: ["--tgz", "/some/path"] });
         });
 
         it('does not blow up when no-tgz arg is at the end', function() {
             // Implicit non-exception
-            hatch = new libingester.Hatch("abad", { argv: ["/blah", "--no-tgz"] });
+            hatch = new libingester.Hatch("abad", "en", { argv: ["/blah", "--no-tgz"] });
         });
 
         it('does not blow up when no-tgz arg is at the end', function() {
             // Implicit non-exception
-            hatch = new libingester.Hatch("abbd", { argv: ["/blah", "--no-tgz"] });
+            hatch = new libingester.Hatch("abbd", "en", { argv: ["/blah", "--no-tgz"] });
         });
 
         it('does not skip tgz by default', function() {
-            hatch = new libingester.Hatch("aaaa", { argv: ["/blah"] });
+            hatch = new libingester.Hatch("aaaa", "en", { argv: ["/blah"] });
             expect(hatch.is_exporting_tgz()).to.be.equal(true);
 
             return hatch.finish().then(() => {
@@ -51,7 +76,7 @@ describe('Hatch', function() {
         });
 
         it('skip tgz if flag set', function() {
-            hatch = new libingester.Hatch("abce", { argv: ["/blah", "--no-tgz"] });
+            hatch = new libingester.Hatch("abce", "en", { argv: ["/blah", "--no-tgz"] });
             expect(hatch.is_exporting_tgz()).to.be.equal(false);
 
             return hatch.finish().then(() => {
@@ -65,16 +90,16 @@ describe('Hatch', function() {
     describe('argv path option', function() {
         it('does not blow up when path arg is not there', function() {
             // Implicit non-exception
-            hatch = new libingester.Hatch("abcd", { argv: ["--foo", "/some/path"] });
+            hatch = new libingester.Hatch("abcd", "en", { argv: ["--foo", "/some/path"] });
         });
 
         it('can process path correctly from passed in argv', function() {
-            hatch = new libingester.Hatch("abcd", { argv: ["--path", "./hatch_foo"] });
+            hatch = new libingester.Hatch("abcd", "en", { argv: ["--path", "./hatch_foo"] });
             expect(hatch.get_path()).to.equal("./hatch_foo");
         });
 
         it('does not break if invalid arg position', function() {
-            hatch = new libingester.Hatch("abcd", { argv: ["foo", "--path"] });
+            hatch = new libingester.Hatch("abcd", "en", { argv: ["foo", "--path"] });
             expect(hatch.get_path()).to.match(/hatch_abcd_[0-9_]+/);
         });
 
@@ -86,7 +111,7 @@ describe('Hatch', function() {
 
             expect(fs.existsSync(targetDir)).to.be.equal(false);
 
-            hatch = new libingester.Hatch("abcd", { argv: ["--path", targetDir] });
+            hatch = new libingester.Hatch("abcd", "en", { argv: ["--path", targetDir] });
             expect(fs.lstatSync(targetDir).isDirectory()).to.be.equal(true);
         });
 
@@ -99,7 +124,7 @@ describe('Hatch', function() {
 
             expect(fs.lstatSync(targetDir).isDirectory()).to.be.equal(true);
 
-            hatch = new libingester.Hatch("abcd", { argv: ["--path", targetDir] });
+            hatch = new libingester.Hatch("abcd", "en", { argv: ["--path", targetDir] });
         });
     });
 });
